add new function: query image using keyword to unsplash
