+++ cloud-functions/functions/index.js
@@ -2,8 +2,10 @@ const functions = require('firebase-functions');
 
 const app = require('express')('cors');
 const cors = require('cors');
+const {searchImageByKeyword} = require('./unsplash');
+const {getImageLabels} = require('./cloud-vision/index');
+
 app.use(cors());
 
-const {getImageLabels} = require('./vision-functions/index');
-
 exports.getImageLabels = functions.region('asia-northeast1').https.onRequest(getImageLabels);
+exports.searchImageByKeyword = functions.region('asia-northeast1').https.onRequest(searchImageByKeyword);
