+++ cloud-functions/functions/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@google-cloud/vision": "^1.11.0",
+    "axios": "^0.19.2",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "firebase-admin": "^8.6.0",
